+++ src/main/java/game/Game.java
@@ -1,5 +1,7 @@
 package game;
 
+import game.log.Logger;
+
 import java.util.ArrayList;
 
 public class Game {
