+++ src/main/java/game/MyKeyListener.java
@@ -1,5 +1,9 @@
 package game;
 
+import game.log.LogSettings;
+import game.log.Logger;
+import game.screens.LogScreen;
+
 import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
 import java.io.IOException;
