+++ src/test/java/BubbleTest.java
@@ -6,10 +6,10 @@ import static org.junit.Assert.*;
 import java.util.ArrayList;
 import java.awt.Color;
 
-import game.Bubble;
-import game.Bubblex16;
-import game.Bubblex8;
 import game.Settings;
+import game.bubble.Bubble;
+import game.bubble.Bubblex16;
+import game.bubble.Bubblex8;
 
 import org.junit.Before;
 import org.junit.Test;
