+++ src/test/java/LogScreenTest.java
@@ -1,7 +1,7 @@
 
 
 import static org.junit.Assert.*;
-import game.LogScreen;
+import game.screens.LogScreen;
 
 import java.io.IOException;
