+++ src/test/java/PlayerBubbleCollisionTest.java
@@ -1,9 +1,9 @@
 import static org.junit.Assert.*;
-import game.Bubble;
-import game.Bubblex16;
 import game.Level;
 import game.Player;
 import game.Rope;
+import game.bubble.Bubble;
+import game.bubble.Bubblex16;
 
 import java.util.ArrayList;
 import java.util.Arrays;
