+++ src/test/java/RopeTest.java
@@ -3,11 +3,11 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
-import game.Bubble;
 import game.Level;
 import game.Player;
 import game.Powerup;
 import game.Rope;
+import game.bubble.Bubble;
 
 import java.util.ArrayList;
